64 bit support
git-svn-id: file:///home/unikey/trunk@113 e1ceb1ec-800a-4d3f-98f3-0cab0dcde3af

+++ src/byteio/byteio.cpp
@@ -516,7 +516,7 @@ int FileBOStream::puts(const char *s, int size)
 		m_bad = (fputs(s, m_file) == EOF);
 		return (!m_bad);
 	}
-	int out = fwrite(s, 1, size, m_file);
+	int out = (int)fwrite(s, 1, size, m_file);
 	m_bad = (out != size);
 	return (!m_bad);
 }
